clientlibs/python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vectron-client"
version = "0.1.0"
authors = [
  { name="Vectron AI", email="contact@vectron.ai" },
]
description = "Official Python client for the Vectron vector database."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "grpcio>=1.60.0",
    "protobuf>=4.0.0",
    "google-api-python-client>=2.0.0",
]

[project.urls]
"Homepage" = "https://github.com/pavandhadge/vectron"
"Bug Tracker" = "https://github.com/pavandhadge/vectron/issues"
